+++ pub/unrouter/lib/unrouter.dart
@@ -6,6 +6,7 @@ library;
 
 export 'src/core/route_data.dart';
 export 'src/core/route_definition.dart';
+export 'src/runtime/adapter_runtime.dart';
 export 'src/runtime/unrouter.dart';
 export 'src/shell/shell.dart';
 export 'src/runtime/runtime_state.dart'